Moved command globals to lib/commands.js
