+++ lib/fontforge.js
@@ -1,10 +1,9 @@
 'use strict';
 
-var which = require('which').sync;
 var execSync = require('child_process').execSync;
 var weightTable = require('./weight.js');
 
-var fontForgeCommand = which('fontforge')
+var fontForgeCommand = require('./commands.js').fontforge;
 
 function FontForgeException(e, cmd) {
    this.message = 'FontForge command failed: ' + e.toString() + '\n' +
