+++ lib/ttf2eot.js
@@ -1,20 +1,20 @@
 'use strict';
 
-var which = require('which').sync;
 var quote = require('./helpers.js').quote;
 var trim = require('./helpers.js').trim;
 var execSync = require('child_process').execSync;
+
 var FontFaceException = require('./exception.js');
 
-var ttf2eotCommand = which('ttf2eot');
+var ttf2eotCommand = require('./commands.js').ttf2eot;
 
-function ttf2eot(source, dest) {
+module.exports = function(source, dest) {
 
     var command = [ttf2eotCommand, quote(source), '>', quote(dest)].join(' ');
 
     try {
 
-      var result = execSync(command);
+        var result = execSync(command).toString();
 
     } catch (e) {
 
@@ -26,6 +26,4 @@ function ttf2eot(source, dest) {
     }
 
     return result;
-}
-
-module.exports = ttf2eot;
+};
