+++ lib/ttf2svg.js
@@ -5,7 +5,7 @@ var quote = require('./helpers.js').quote;
 var trim = require('./helpers.js').trim;
 var isLinux = require('./helpers.js').isLinux();
 
-var ttf2svgCommand = (isLinux ? 'ttf2svg': 'batik-ttf2svg');
+var ttf2svgCommand = require('./commands.js').ttf2svg;
 
 function ttf2svg(source, target, name) {
